rename var_cov to moment
