+++ crates/polars-compute/src/rolling/nulls/variance.rs
@@ -2,7 +2,7 @@
 use num_traits::{FromPrimitive, ToPrimitive};
 
 use super::*;
-use crate::var_cov::VarState;
+use crate::moment::VarState;
 
 pub struct VarWindow<'a, T> {
     slice: &'a [T],
